refactored lucidGateway file into classes
