+++ editorextensions/codebeamer-custom-ui/modal/src/api/lucidGateway.ts
@@ -1,188 +1,11 @@
-import { CodeBeamerItem } from '../models/codebeamer-item.if';
-import getItemColorField from './utils/getItemColorField';
-import { store } from '../store/store';
-import TrackerDetails from '../models/trackerDetails.if';
-import { CardData } from '../models/lucidCardData';
-import { RelationshipType } from '../enums/associationRelationshipType.enum';
-import { getColorForRelationshipType } from './utils/getColorForRelationshipType';
-
-/**
- * Class for handling message events and callbacks.
- */
-export class MessageHandler {
-	private callbacks: Map<MessageAction, ((data: []) => void)[]> = new Map();
-
-	/**
-	 * Private instance to hold the singleton instance.
-	 * @type {MessageHandler | null}
-	 * @private
-	 */
-	private static instance: MessageHandler | null = null;
-
-	/**
-	 * Private constructor to prevent direct instantiation. Sets up a message event listener.
-	 * @private
-	 */
-	private constructor() {
-		window.addEventListener('message', (e) => {
-			const data = JSON.parse(e.data);
-			this.notifyCallbacks(data);
-		});
-	}
-
-	/**
-	 * Gets the singleton instance of the MessageHandler.
-	 * If the instance doesn't exist, it creates one.
-	 * @returns {MessageHandler} The singleton instance of MessageHandler.
-	 * @static
-	 */
-	public static getInstance(): MessageHandler {
-		if (!MessageHandler.instance) {
-			MessageHandler.instance = new MessageHandler();
-		}
-		return MessageHandler.instance;
-	}
-
-	/**
-	 * Requests card blocks from the parent window and registers a callback to handle the response.
-	 * @param {function} callback - The callback function that will be called with the received card block data.
-	 */
-	getCardBlocks(callback: (data: CardBlockData[]) => void) {
-		this.subscribeCallback(MessageAction.GET_CARD_BLOCKS, callback);
-		LucidGateway.requestCardBlockData();
-	}
-
-	/**
-	 * Requests lines from the parent window and registers a callback to handle the response.
-	 * @param {function} callback - The callback function that will be called with the received line data.
-	 */
-	getLines(callback: (data: LineData[]) => void) {
-		this.subscribeCallback(MessageAction.GET_LINES, callback);
-		LucidGateway.requestLineData();
-	}
-
-	subscribeCallback(action: MessageAction, callback: (data: []) => void) {
-		const actionCallbacks = this.callbacks.get(action) || [];
-		actionCallbacks.push(callback);
-		this.callbacks.set(action, actionCallbacks);
-	}
-
-	unsubscribeCallback(action: MessageAction, callback: (data: []) => void) {
-		const actionCallbacks = this.callbacks.get(action) || [];
-		const index = actionCallbacks.indexOf(callback);
-		if (index !== -1) {
-			actionCallbacks.splice(index, 1);
-		}
-		this.callbacks.set(action, actionCallbacks);
-	}
-
-	/**
-	 * Notifies registered callbacks with message data.
-	 * @param data - The data received in the message.
-	 */
-	private notifyCallbacks(data: IncomingMessage) {
-		const actionCallbacks = this.callbacks.get(data.action);
-		if (actionCallbacks) {
-			actionCallbacks.forEach((callback) => {
-				callback(data.payload as []);
-				this.unsubscribeCallback(data.action, callback);
-			});
-		}
-	}
-}
-
-/**
- * Interface for a message object with an action and payload.
- * @interface Message
- * @property action - The action to perform.
- * @property payload - The payload for the action.
- */
-export interface Message {
-	action: MessageAction;
-	payload?:
-		| ImportItemPayload
-		| UpdateCardPayload
-		| StartImportPayload
-		| CreateLinePayload
-		| DeleteLinePayload;
-}
-
-export interface IncomingMessage {
-	action: MessageAction;
-	payload: LineData[] | CardBlockData[];
-}
-
-// Interface for payload specific to the IMPORT_ITEM action
-export interface ImportItemPayload {
-	importId: number;
-	cardData: CardData;
-}
-
-// Interface for payload specific to the UPDATE_CARD action
-export interface UpdateCardPayload {
-	cardData: CardData;
-	cardBlockId: string;
-}
-
-// Interface for payload specific to the CREATE_LINE action
-export interface CreateLinePayload {
-	sourceBlockId: string;
-	targetBlockId: string;
-	relationshipType: RelationshipType;
-	lineColor: string;
-}
-
-// Interface for payload specific to the DELETE_LINE action
-export interface DeleteLinePayload {
-	lineId: string;
-}
-
-// Interface for payload specific to the START_IMPORT action
-export interface StartImportPayload {
-	id: number;
-	totalItems: number;
-}
-
-/**
- * action enum for message actions
- */
-export enum MessageAction {
-	GET_CARD_BLOCKS = 'getCardBlocks',
-	GET_LINES = 'getLines',
-	IMPORT_ITEM = 'importItem',
-	UPDATE_CARD = 'updateCard',
-	CREATE_LINE = 'createLine',
-	DELETE_LINE = 'deleteLine',
-	START_IMPORT = 'startImport',
-	CLOSE_MODAL = 'closeModal',
-}
-
-/**
- * Interface for card block data.
- */
-export interface CardBlockData {
-	cardBlockId: string;
-	codebeamerItemId: number;
-	codebeamerTrackerId: number;
-}
-
-/**
- * Interface for line data
- */
-export interface LineData {
-	id: string;
-	sourceBlockId: string;
-	targetBlockId: string;
-}
-
-/**
- * Structure of a Relation with the specific block ids
- */
-export interface BlockRelation {
-	sourceBlockId: string;
-	targetBlockId: string;
-	type: RelationshipType;
-}
+import { CodeBeamerItem } from "../models/codebeamer-item.if";
+import getItemColorField from "./utils/getItemColorField";
+import { store } from "../store/store";
+import TrackerDetails from "../models/trackerDetails.if";
+import { CardData } from "../models/lucidCardData";
+import { RelationshipType } from "../enums/associationRelationshipType.enum";
+import { getColorForRelationshipType } from "./utils/getColorForRelationshipType";
+import { Message, MessageAction } from "../models/messageInterfaces";
 
 export class LucidGateway {
 	/**
@@ -259,8 +82,8 @@ export class LucidGateway {
 		const cbBaseAddress = store.getState().boardSettings.cbAddress;
 
 		const headers = new Headers({
-			'Content-Type': 'text/plain',
-			Authorization: `Basic ${btoa(username + ':' + password)}`,
+			"Content-Type": "text/plain",
+			Authorization: `Basic ${btoa(username + ":" + password)}`,
 		});
 
 		// if (item.descriptionFormat == DescriptionFormat.WIKI) {
@@ -285,7 +108,7 @@ export class LucidGateway {
 		try {
 			const trackerRes = await fetch(
 				`${cbBaseAddress}/api/v3/trackers/${item.tracker.id}`,
-				{ method: 'GET', headers }
+				{ method: "GET", headers }
 			);
 			const trackerJson = (await trackerRes.json()) as TrackerDetails;
 			item.tracker.keyName = trackerJson.keyName;
@@ -377,6 +200,6 @@ export class LucidGateway {
 	 * Post a message to the parent window.
 	 */
 	private static postMessage(message: Message) {
-		window.parent.postMessage(message, '*');
+		window.parent.postMessage(message, "*");
 	}
 }
