+++ editorextensions/codebeamer-custom-ui/modal/src/pages/import/components/updater/Updater.tsx
@@ -1,20 +1,20 @@
-import React, { useState } from 'react';
-import { Modal, Spinner, ProgressBar } from 'react-bootstrap';
-import { useDispatch, useSelector } from 'react-redux';
+import React, { useState } from "react";
+import { Modal, Spinner, ProgressBar } from "react-bootstrap";
+import { useDispatch, useSelector } from "react-redux";
 import {
 	useGetItemsQuery,
 	useGetTrackerDetailsQuery,
-} from '../../../../api/codeBeamerApi';
-import { LucidGateway } from '../../../../api/lucidGateway';
+} from "../../../../api/codeBeamerApi";
+import { LucidGateway } from "../../../../api/LucidGateway";
 import {
 	DEFAULT_RESULT_PAGE,
 	MAX_ITEMS_PER_SYNCH,
-} from '../../../../constants/cb-import-defaults';
-import { CardBlockToItemMapping } from '../../../../models/cardBlockToItemMapping.if';
-import { CodeBeamerItem } from '../../../../models/codebeamer-item.if';
-import { RootState } from '../../../../store/store';
+} from "../../../../constants/cb-import-defaults";
+import { CardBlockToItemMapping } from "../../../../models/cardBlockToItemMapping.if";
+import { CodeBeamerItem } from "../../../../models/codebeamer-item.if";
+import { RootState } from "../../../../store/store";
 
-import '../importer/importer.css';
+import "../importer/importer.css";
 
 /**
  * Twin of {@link Importer}, but for updating.
@@ -38,9 +38,7 @@ export default function Updater(props: {
 	const { data, error, isLoading } = useGetItemsQuery({
 		page: DEFAULT_RESULT_PAGE,
 		pageSize: MAX_ITEMS_PER_SYNCH,
-		queryString: `item.id IN (${props.items
-			.map((i) => i.itemId)
-			.join(',')})`,
+		queryString: `item.id IN (${props.items.map((i) => i.itemId).join(",")})`,
 	});
 
 	const {
@@ -64,7 +62,7 @@ export default function Updater(props: {
 				if (_items[i].categories?.length) {
 					if (
 						_items[i].categories.find(
-							(c) => c.name == 'Folder' || c.name == 'Information'
+							(c) => c.name == "Folder" || c.name == "Information"
 						)
 					) {
 						// miro.board.notifications.showInfo(
